Refactor types to imported Zod schema
